Update README.md
The implementation and results of a PageRank-based iterative method to identify the most important people occurring in news articles. The data used is a plain text file containing an undirected and unweighted graph of a social network of co-occurrence in news articles. The graph has been constructed from a subset of 3000 news articles from the Reuters-21578 corpus by identifying person names.

Methodology: 

The PageRank algorithm is implemented from scratch using Python, with the power iteration method. The provided network was undirected, and therefore, before applying the PageRank algorithm, it was first converted to a directed network. 

a)	For each edge vertex1 vertex2, an edge in the opposite direction, i.e., vertex2 vertex1, was included. 
b)	The teleportation rate was set to 0.10.